+++ core/src/main/java/org/jdbi/v3/core/statement/SqlStatements.java
@@ -29,8 +29,10 @@ import java.util.function.Function;
 
 import javax.annotation.Nullable;
 
-import com.github.benmanes.caffeine.cache.Cache;
-import com.github.benmanes.caffeine.cache.Caffeine;
+import org.jdbi.v3.core.cache.JdbiCache;
+import org.jdbi.v3.core.cache.JdbiCacheBuilder;
+import org.jdbi.v3.core.cache.JdbiCacheLoader;
+import org.jdbi.v3.core.cache.internal.DefaultJdbiCacheBuilder;
 import org.jdbi.v3.core.config.JdbiConfig;
 import org.jdbi.v3.meta.Beta;
 
@@ -39,9 +41,12 @@ import org.jdbi.v3.meta.Beta;
  */
 public final class SqlStatements implements JdbiConfig<SqlStatements> {
 
+    /** The default size of the SQL template cache. */
+    public static final int SQL_TEMPLATE_CACHE_SIZE = 1_000;
+
     private final Map<String, Object> attributes;
     private TemplateEngine templateEngine;
-    private Cache<StatementCacheKey, Function<StatementContext, String>> templateCache;
+    private JdbiCache<StatementCacheKey, Function<StatementContext, String>> templateCache;
     private SqlParser sqlParser;
     private SqlLogger sqlLogger;
     private Integer queryTimeout;
@@ -56,7 +61,7 @@ public final class SqlStatements implements JdbiConfig<SqlStatements> {
         sqlParser = new ColonPrefixSqlParser();
         sqlLogger = SqlLogger.NOP_SQL_LOGGER;
         queryTimeout = null;
-        templateCache = Caffeine.newBuilder().maximumSize(1_000).build();
+        templateCache = DefaultJdbiCacheBuilder.builder().maxSize(SQL_TEMPLATE_CACHE_SIZE).build();
     }
 
     private SqlStatements(SqlStatements that) {
@@ -158,14 +163,14 @@ public final class SqlStatements implements JdbiConfig<SqlStatements> {
     }
 
     /**
-     * Sets the Caffeine cache used to avoid repeatedly parsing SQL statements.
+     * Sets the cache used to avoid repeatedly parsing SQL statements.
      *
-     * @param caffeineSpec the cache builder to use to cache parsed SQL
+     * @param cacheBuilder the cache builder to use to create the cache.
      * @return this
      */
     @Beta
-    public SqlStatements setTemplateCache(Caffeine<Object, Object> caffeineSpec) {
-        templateCache = caffeineSpec.build();
+    public SqlStatements setTemplateCache(JdbiCacheBuilder cacheBuilder) {
+        templateCache = cacheBuilder.build();
         return this;
     }
 
@@ -263,6 +268,16 @@ public final class SqlStatements implements JdbiConfig<SqlStatements> {
         return this;
     }
 
+    /**
+     * Returns cache statistics for the internal template cache. This returns a cache specific object,
+     * so the user needs to know what caching library is in use.
+     */
+    @Beta
+    public <T> T cacheStats() {
+        return templateCache.getStats();
+    }
+
+
     void customize(Statement statement) throws SQLException {
         if (queryTimeout != null) {
             statement.setQueryTimeout(queryTimeout);
@@ -285,15 +300,18 @@ public final class SqlStatements implements JdbiConfig<SqlStatements> {
     String preparedRender(String template, StatementContext ctx) {
         try {
             return Optional.ofNullable(
-                            templateCache.get(
+                            templateCache.getWithLoader(
                                     new StatementCacheKey(templateEngine, template),
-                                    key -> key.getTemplateEngine().parse(key.getTemplate(), ctx.getConfig())
-                                            .orElse(null))) // no parse -> no cache
+                                    cacheLoaderFunction(ctx)))
                     .orElse(cx -> templateEngine.render(template, cx)) // fall-back to old behavior
                     .apply(ctx);
         } catch (final IllegalArgumentException e) {
             throw new UnableToCreateStatementException("Exception rendering SQL template", e, ctx);
         }
+    }
+
+    private static JdbiCacheLoader<StatementCacheKey, Function<StatementContext, String>> cacheLoaderFunction(StatementContext ctx) {
+        return key -> key.getTemplateEngine().parse(key.getTemplate(), ctx.getConfig()).orElse(null);
     }
 
     private static final class StatementCacheKey {